Refactor: Utils/ dir with sub-dir created
