+++ financial_loss_functions/scripts/run_training_grid.py
@@ -2,7 +2,7 @@ import os
 import sys
 import signal
 import argparse
-from src.utils import load_path_config, load_config
+from src.utils.io import load_path_config, load_config
 from src.training.pipeline import run_training_pipeline
 
 _interrupted = False
