+++ financial_loss_functions/src/utils/io.py
@@ -1,6 +1,5 @@
 import os
 import json
-import torch
 import shutil
 import pandas as pd
 from pathlib import Path
@@ -156,30 +155,3 @@ def load_config(path: str) -> dict:
     with open(path, 'r') as f:
         config = json.load(f)
     return config
-
-
-def get_best_device() -> torch.device:
-    if torch.backends.mps.is_available():
-        print('Using mps for GPU acceleration.')
-        return torch.device('mps')
-    
-    elif torch.cuda.is_available():
-        print('Using cuda for GPU acceleration.')
-        return torch.device('cuda')
-    
-    else:
-        print('No GPU acceleration. Using CPU.')
-        return torch.device('cpu')
-
-def extract_req_cols(columns_list: list, suffix: str) -> list:
-    """
-    Extract required columns based on the suffix in the column names. e.g., NSDN_RETURN
-
-    @param columns_list list List of all column names.
-    @param suffix str 
-        Suffix str to extract its respective columns. e.g., VOL_CHANGE, RETURN
-    
-    @return required_cols List of required column names for the given suffix
-    """
-    required_cols = [col for col in columns_list if suffix in col]
-    return required_cols
